Remove javax.servlet dependency from Core. Splitting OSLC4JUtils into ProviderHelper class, for functions that are needed by Providers. Moving ProviderHelper class from Jena-provider to Utils project. Making "private" methods public in Provider Helper.
