+++ core/oslc4j-core/src/test/java/org/eclipse/lyo/oslc4j/core/OSLC4JUtilsTest.java
@@ -14,40 +14,10 @@
 package org.eclipse.lyo.oslc4j.core;
 
 import java.net.MalformedURLException;
-import javax.servlet.http.HttpServletRequest;
 import static org.junit.Assert.*;
-import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
-import static org.mockito.Mockito.when;
 
 public class OSLC4JUtilsTest {
-
-    @Before
-    public void clearPublicURISystemProperty() throws MalformedURLException {
-        System.setProperty(OSLC4JConstants.OSLC4J_DISABLE_HOST_RESOLUTION, "true");
-        OSLC4JUtils.setPublicURI(null);
-        OSLC4JUtils.setServletPath(null);
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testPublicURI() throws MalformedURLException {
-        HttpServletRequest mockedRequest = Mockito.mock(HttpServletRequest.class);
-        when(mockedRequest.getServletPath()).thenReturn("/resources");
-        when(mockedRequest.getPathInfo()).thenReturn("/bugs/1");
-        OSLC4JUtils.setPublicURI("http://hostname.example.com:12357/myapp/");
-        String resolvedUri = OSLC4JUtils.resolveURI(mockedRequest, true);
-        assertEquals("http://hostname.example.com:12357/myapp/resources/bugs/1", resolvedUri);
-    }
-
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testDisableHostResolution() {
-        HttpServletRequest mockedRequest = mockRequest();
-        String resolvedUri = OSLC4JUtils.resolveURI(mockedRequest, true);
-        assertEquals("https://hostname.example.com:12357/myapp/resources/bugs/1", resolvedUri);
-    }
 
     @Test(expected = IllegalArgumentException.class)
     public void setServletPathNull() throws Exception {
@@ -73,46 +43,6 @@ public class OSLC4JUtilsTest {
         final String servletURI = OSLC4JUtils.getServletURI();
 
         assertEquals("http://a.b/abc", servletURI);
-    }
-
-    @Test
-    public void resolveFullPathUriWithPublicUri() throws Exception {
-        OSLC4JUtils.setPublicURI("http://a.b/");
-        OSLC4JUtils.setServletPath("/abc");
-        final HttpServletRequest request = mockRequest();
-
-        final String fullUri = OSLC4JUtils.resolveFullUri(request);
-
-        assertEquals("http://a.b/abc/bugs/1", fullUri);
-    }
-
-    @Test
-    public void resolveFullPathUriWithoutPublicUri() {
-        final HttpServletRequest request = mockRequest();
-
-        final String fullUri = OSLC4JUtils.resolveFullUri(request);
-
-        assertEquals("https://hostname.example.com:12357/myapp/resources/bugs/1", fullUri);
-    }
-
-    @Test
-    public void resolveServletUriWithPublicUri() throws Exception {
-        OSLC4JUtils.setPublicURI("http://a.b/");
-        OSLC4JUtils.setServletPath("/abc");
-        final HttpServletRequest request = mockRequest();
-
-        final String fullUri = OSLC4JUtils.resolveServletUri(request);
-
-        assertEquals("http://a.b/abc", fullUri);
-    }
-
-    @Test
-    public void resolveServletUriWithoutPublicUri() {
-        final HttpServletRequest request = mockRequest();
-
-        final String fullUri = OSLC4JUtils.resolveServletUri(request);
-
-        assertEquals("https://hostname.example.com:12357/myapp/resources", fullUri);
     }
 
     @Test
@@ -164,14 +94,4 @@ public class OSLC4JUtilsTest {
         assertFalse(OSLC4JUtils.hasLyoStorePagingPreciseLimit());
     }
 
-    private HttpServletRequest mockRequest() {
-        HttpServletRequest mockedRequest = Mockito.mock(HttpServletRequest.class);
-        when(mockedRequest.getScheme()).thenReturn("https");
-        when(mockedRequest.getServerName()).thenReturn("hostname.example.com");
-        when(mockedRequest.getServerPort()).thenReturn(12357);
-        when(mockedRequest.getContextPath()).thenReturn("/myapp");
-        when(mockedRequest.getServletPath()).thenReturn("/resources");
-        when(mockedRequest.getPathInfo()).thenReturn("/bugs/1");
-        return mockedRequest;
-    }
 }
