+++ server/oslc4j-jena-provider/src/main/java/org/eclipse/lyo/oslc4j/provider/jena/OslcRdfXmlProvider.java
@@ -39,6 +39,7 @@ import org.eclipse.lyo.oslc4j.core.model.OslcMediaType;
 import org.eclipse.lyo.oslc4j.core.model.ResponseInfo;
 import org.eclipse.lyo.oslc4j.core.model.ResponseInfoArray;
 import org.eclipse.lyo.oslc4j.core.model.ResponseInfoCollection;
+import org.eclipse.lyo.oslc4j.provider.helper.ProviderHelper;
 
 /**
  * @author Russell Boykin, Alberto Giammaria, Chris Peters, Gianluca Bernardini, Steve Pitschke
@@ -153,7 +154,7 @@ public class OslcRdfXmlProvider
 			{
 				responseInfo = (ResponseInfo<?>)filteredResource;
 				
-				String requestURI = OSLC4JUtils.resolveURI(httpServletRequest, true);
+				String requestURI = ProviderHelper.resolveURI(httpServletRequest, true);
 				responseInfoURI = requestURI;
 				
 				FilteredResource<?> container = responseInfo.getContainer();
