+++ server/oslc4j-provider-utils/src/test/java/org/eclipse/lyo/oslc4j/provider/helper/OSLC4JUtilsTest.java
@@ -1,0 +1,106 @@
+/*
+ * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Eclipse Distribution License 1.0
+ * which is available at http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+ */
+package org.eclipse.lyo.oslc4j.provider.helper;
+
+import java.net.MalformedURLException;
+import javax.servlet.http.HttpServletRequest;
+import static org.junit.Assert.*;
+
+import org.eclipse.lyo.oslc4j.core.OSLC4JConstants;
+import org.eclipse.lyo.oslc4j.core.OSLC4JUtils;
+import org.eclipse.lyo.oslc4j.provider.helper.ProviderHelper;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mockito;
+import static org.mockito.Mockito.when;
+
+public class OSLC4JUtilsTest {
+
+    @Before
+    public void clearPublicURISystemProperty() throws MalformedURLException {
+        System.setProperty(OSLC4JConstants.OSLC4J_DISABLE_HOST_RESOLUTION, "true");
+        OSLC4JUtils.setPublicURI(null);
+        OSLC4JUtils.setServletPath(null);
+    }
+
+    @Test
+    @SuppressWarnings("deprecation")
+    public void testPublicURI() throws MalformedURLException {
+        HttpServletRequest mockedRequest = Mockito.mock(HttpServletRequest.class);
+        when(mockedRequest.getServletPath()).thenReturn("/resources");
+        when(mockedRequest.getPathInfo()).thenReturn("/bugs/1");
+        OSLC4JUtils.setPublicURI("http://hostname.example.com:12357/myapp/");
+        String resolvedUri = ProviderHelper.resolveURI(mockedRequest, true);
+        assertEquals("http://hostname.example.com:12357/myapp/resources/bugs/1", resolvedUri);
+    }
+
+    @Test
+    @SuppressWarnings("deprecation")
+    public void testDisableHostResolution() {
+        HttpServletRequest mockedRequest = mockRequest();
+        String resolvedUri = ProviderHelper.resolveURI(mockedRequest, true);
+        assertEquals("https://hostname.example.com:12357/myapp/resources/bugs/1", resolvedUri);
+    }
+
+    @Test
+    public void resolveFullPathUriWithPublicUri() throws Exception {
+        OSLC4JUtils.setPublicURI("http://a.b/");
+        OSLC4JUtils.setServletPath("/abc");
+        final HttpServletRequest request = mockRequest();
+
+        final String fullUri = ProviderHelper.resolveFullUri(request);
+
+        assertEquals("http://a.b/abc/bugs/1", fullUri);
+    }
+
+    @Test
+    public void resolveFullPathUriWithoutPublicUri() {
+        final HttpServletRequest request = mockRequest();
+
+        final String fullUri = ProviderHelper.resolveFullUri(request);
+
+        assertEquals("https://hostname.example.com:12357/myapp/resources/bugs/1", fullUri);
+    }
+
+    @Test
+    public void resolveServletUriWithPublicUri() throws Exception {
+        OSLC4JUtils.setPublicURI("http://a.b/");
+        OSLC4JUtils.setServletPath("/abc");
+        final HttpServletRequest request = mockRequest();
+
+        final String fullUri = ProviderHelper.resolveServletUri(request);
+
+        assertEquals("http://a.b/abc", fullUri);
+    }
+
+    @Test
+    public void resolveServletUriWithoutPublicUri() {
+        final HttpServletRequest request = mockRequest();
+
+        final String fullUri = ProviderHelper.resolveServletUri(request);
+
+        assertEquals("https://hostname.example.com:12357/myapp/resources", fullUri);
+    }
+
+    private HttpServletRequest mockRequest() {
+        HttpServletRequest mockedRequest = Mockito.mock(HttpServletRequest.class);
+        when(mockedRequest.getScheme()).thenReturn("https");
+        when(mockedRequest.getServerName()).thenReturn("hostname.example.com");
+        when(mockedRequest.getServerPort()).thenReturn(12357);
+        when(mockedRequest.getContextPath()).thenReturn("/myapp");
+        when(mockedRequest.getServletPath()).thenReturn("/resources");
+        when(mockedRequest.getPathInfo()).thenReturn("/bugs/1");
+        return mockedRequest;
+    }
+}
